fix: move iCalendar generation from endpoint to Application handler
RFC 5545 ICS formatting (VEVENT construction, escaping, line folding)
lived entirely in GetEngagementCalendarEndpoint instead of an
Application-layer handler, violating this repo's Clean Architecture
convention. Moved the formatting into GetEngagementCalendarQueryHandler,
which now returns a fully-rendered EngagementCalendarFile; the endpoint
only resolves the CORS-derived base URL and wraps the result as a file
response.

Fixes #876
