+++ backend/src/Application/Engagements/EngagementCalendarFile.cs
@@ -1,0 +1,3 @@
+namespace Application.Engagements;
+
+public sealed record EngagementCalendarFile(string FileName, string Content);
