+++ backend/src/Application/Engagements/GetEngagementCalendar/v1/GetEngagementCalendarQuery.cs
@@ -1,0 +1,6 @@
+using Application.Common.Messaging;
+
+namespace Application.Engagements.GetEngagementCalendar.v1;
+
+public sealed record GetEngagementCalendarQuery(Guid EngagementId, string BaseUrl)
+	: IQuery<EngagementCalendarFile?>;
