--- backend/src/Application/Engagements/GetEngagementCalendarInfo/v1/GetEngagementCalendarInfoQuery.cs
@@ -1,6 +1,0 @@
-using Application.Common.Messaging;
-
-namespace Application.Engagements.GetEngagementCalendarInfo.v1;
-
-public sealed record GetEngagementCalendarInfoQuery(Guid EngagementId)
-	: IQuery<EngagementCalendarInfo?>;
