--- backend/src/Application/Engagements/GetEngagementCalendarInfo/v1/GetEngagementCalendarInfoQueryHandler.cs
@@ -1,16 +1,0 @@
-using Application.Common.Exceptions;
-using Application.Common.Messaging;
-using Domain.Engagements;
-
-namespace Application.Engagements.GetEngagementCalendarInfo.v1;
-
-internal sealed class GetEngagementCalendarInfoQueryHandler(
-	IEngagementReadRepository readRepository)
-	: IQueryHandler<GetEngagementCalendarInfoQuery, EngagementCalendarInfo?>
-{
-	public async ValueTask<EngagementCalendarInfo?> Handle(
-		GetEngagementCalendarInfoQuery request,
-		CancellationToken cancellationToken = default) =>
-			await readRepository.GetCalendarInfoAsync(
-				EngagementId.Create(request.EngagementId).GetValueOrThrow(), cancellationToken);
-}
